Move the "testing helpers" back into export_test.go.
