+++ state/backups/export_test.go
@@ -3,6 +3,24 @@
 
 package backups
 
+import (
+	"io"
+
+	"github.com/juju/juju/state/backups/db"
+)
+
 var (
 	Create = create
 )
+
+func ExposeCreateResult(result *createResult) (io.ReadCloser, int64, string) {
+	return result.archiveFile, result.size, result.checksum
+}
+
+func NewTestCreateArgs(filesToBackUp []string, db db.Dumper) *createArgs {
+	args := createArgs{
+		filesToBackUp: filesToBackUp,
+		db:            db,
+	}
+	return &args
+}
